moved JSON to kvstore
